requirements.txt
Django==3.2.25
Mezzanine==5.1.4
grappelli-safe==1.1.1
gunicorn==22.0.0
whitenoise==6.6.0
django-storages[google]==1.14.3
python-dotenv==1.0.1